# src/main/resources/api/openapi.yaml
openapi: 3.1.0

info:
  title: OrderService API
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local server url

tags:
  - name: Order

security:
  - BearerAuth: []
  - OAuth2: []

paths:
  /orders/{id}:
    put:
      tags:
        - Order
      security:
        - BearerAuth: [ ]
      operationId: updateOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrder'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
    delete:
      tags:
        - Order
      operationId: deleteOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
    patch:
      tags:
        - Order
      operationId: patchOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrder'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
  /orders:
    get:
      tags:
        - Order
      operationId: getOrders
      parameters:
        - name: minTotalAmount
          in: query
          required: false
          schema:
            type: number
        - name: maxTotalAmount
          in: query
          required: false
          schema:
            type: number
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
    post:
      tags:
        - Order
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrder'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    OAuth2:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: http://localhost:9090/realms/owu-ecommerce/protocol/openid-connect/auth
          scopes:
            offline_access: offline access
  schemas:
    CreateOrderItem:
      type: object
      properties:
        price:
          type: number
          minimum: 0
        quantity:
          type: integer
          format: int32
          minimum: 1
      required:
        - price
        - quantity
    UpdateOrder:
      type: object
      properties:
        status:
          type: string
          minLength: 1
        invoice:
          $ref: '#/components/schemas/UpsertInvoice'
        items:
          type: array
          items:
            $ref: '#/components/schemas/CreateOrderItem'
          maxItems: 10
          minItems: 1
    UpsertInvoice:
      type: object
      properties:
        number:
          type: string
        totalAmount:
          type: number
          minimum: 0
    Order:
      type: object
      properties:
        id:
          type: integer
          format: int64
        status:
          type: string
        orderDate:
          type: string
          format: date-time
        totalAmount:
          type: number
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
    OrderItem:
      type: object
      properties:
        productPrice:
          type: number
        quantity:
          type: integer
          format: int32
    CreateOrder:
      type: object
      properties:
        shopId:
          type: string
          minLength: 1
        items:
          type: array
          items:
            $ref: '#/components/schemas/CreateOrderItem'
          maxItems: 10
          minItems: 1